requirements.txt
mysqlclient
argparse==1.2.1
begins==0.9
funcsigs==0.4
#logging==0.4.9.6
pyserial==2.7
#wsgiref==0.1.2
